Update README.md
# Anomaly-activity-recognition-using-machine-learning-based-on-video-surveillance

Abstract 
In the current era, where one of the most exciting tools that have been popularized
in the toolbox in recent years is machine learning because of advancement in cloud computing.It resulted in a progress of lot of stagnated research we can finally make further advancement in “Anomaly Detection and Activity Recognition of Humans”.
The major drawback in the traditional approach, that there is a need to perform manual operation for detecting anomaly which lead to possibilities of human errors. This paper focuses on “Anomaly Detection and Activity Recognition of Humans”. The anomaly detection system uses Convolutional Neural Network (CNN) to solve the problems of manual operation such as the false alarms, missing of anomalous events and locating the position of an anomaly in the image. The image wise abnormal event is detected and the location of the abnormality in a image is detected using Convolutional Neural Network.